// src/api-client/models/inventory-item-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Lifecycle state.
 * @export
 * @enum {string}
 */

export const InventoryItemStatus = {
    InInventory: 'IN_INVENTORY',
    Reserved: 'RESERVED',
    Donated: 'DONATED',
    Withdrawn: 'WITHDRAWN',
    Expired: 'EXPIRED'
} as const;

export type InventoryItemStatus = typeof InventoryItemStatus[keyof typeof InventoryItemStatus];