// src/api-client/models/surplus-urgency.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Urgency band, derived from the hours remaining.
 * @export
 * @enum {string}
 */

export const SurplusUrgency = {
    Critical: 'CRITICAL',
    Expiring: 'EXPIRING',
    Standard: 'STANDARD'
} as const;

export type SurplusUrgency = typeof SurplusUrgency[keyof typeof SurplusUrgency];